+++ packages/@headlessng-primitives/description/src/lib/description.directive.spec.ts
@@ -2,16 +2,7 @@ import { Component, DebugElement } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 
-import { FieldDirective } from '../../../field/src';
-
 import { DescriptionDirective } from './description.directive';
-
-class FieldMockDirective {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  public registerDescriptionRef(ref: DescriptionDirective): void {
-    return;
-  }
-}
 
 @Component({
   imports: [DescriptionDirective],
@@ -19,14 +10,8 @@ class FieldMockDirective {
   template: `
     <span hDescription #hDescriptionRef="hDescriptionRef"></span>
     <span hDescription #hDescriptionRef="hDescriptionRef"></span>
-    <span hDescription #hDescriptionRef="hDescriptionRef" class="custom" id="custom-id"></span>
-  `,
-  providers: [
-    {
-      provide: FieldDirective,
-      useClass: FieldMockDirective
-    }
-  ]
+    <span hDescription #hDescriptionRef="hDescriptionRef"></span>
+  `
 })
 class DescriptionSpecComponent {}
 
@@ -45,20 +30,10 @@ describe('@headlessng/primitives/description', () => {
 
     // This test should be run first due to the generation of unique prefixes that are stored in memory.
     it('should generate correct "id" attribute for each element when not passed in input', () => {
-      debugs
-        .map(x => x.nativeElement as HTMLSpanElement)
-        .filter(x => !x.classList.contains('custom'))
-        .forEach((x, i) => {
-          fixture.detectChanges();
-          expect(x.getAttribute('id')).toBe(`h-description-${i}`);
-        });
-    });
-
-    it('should set correct "id" attribute when they are passed to inputs', () => {
-      debugs
-        .map(x => x.nativeElement as HTMLSpanElement)
-        .filter(x => x.classList.contains('custom'))
-        .forEach(x => expect(x.getAttribute('id')).toBe('custom-id'));
+      debugs.forEach((x, i) => {
+        fixture.detectChanges();
+        expect(x.nativeElement.getAttribute('id')).toBe(`h-description-${i}`);
+      });
     });
 
     it('should render the correct description elements', () => {
@@ -69,16 +44,6 @@ describe('@headlessng/primitives/description', () => {
       debugs.forEach(x =>
         expect(x.references['hDescriptionRef']).toBeInstanceOf(DescriptionDirective)
       );
-    });
-
-    it('should correctly register own reference in the field directive', () => {
-      debugs
-        .map(x => x.references['hDescriptionRef'] as DescriptionDirective)
-        .forEach(x => {
-          const registerFn = jest.spyOn(x['_fieldRef'] as FieldDirective, 'registerDescriptionRef');
-          fixture.detectChanges();
-          expect(registerFn).toHaveBeenCalledWith(x);
-        });
     });
   });
 });