+++ packages/@headlessng-primitives/description/src/lib/description.directive.ts
@@ -1,10 +1,5 @@
-import { Directive, effect, ElementRef, inject, Injector, input } from '@angular/core';
-import { FieldDirective, FieldElementRef } from '@headlessng/primitives/field';
-
-const getId = (() => {
-  let id = 0;
-  return () => `h-description-${id++}`;
-})();
+import { Directive } from '@angular/core';
+import { DescriptionFieldRef } from '@headlessng/primitives/field';
 
 @Directive({
   exportAs: 'hDescriptionRef',
@@ -14,19 +9,4 @@ const getId = (() => {
   selector: '[hDescription]',
   standalone: true
 })
-export class DescriptionDirective implements FieldElementRef {
-  private readonly _injector = inject(Injector);
-  private readonly _fieldRef = inject(FieldDirective, { optional: true });
-  private readonly _fieldRefEffect = effect(
-    () => {
-      this._fieldRef?.registerDescriptionRef(this);
-    },
-    {
-      allowSignalWrites: true,
-      injector: this._injector
-    }
-  );
-
-  public readonly elementRef = inject<ElementRef<HTMLElement>>(ElementRef);
-  public readonly id = input<string>(getId(), { alias: 'id' });
-}
+export class DescriptionDirective extends DescriptionFieldRef {}