+++ packages/@headlessng-primitives/field/src/lib/index.ts
@@ -1,1 +1,2 @@
+export * from './field-ref';
 export * from './field.directive';