+++ packages/@headlessng-primitives/label/src/lib/label.directive.spec.ts
@@ -2,16 +2,7 @@ import { Component, DebugElement } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 
-import { FieldDirective } from '../../../field/src';
-
 import { LabelDirective } from './label.directive';
-
-class FieldMockDirective {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  public registerLabelRef(ref: LabelDirective): void {
-    return;
-  }
-}
 
 @Component({
   imports: [LabelDirective],
@@ -19,14 +10,8 @@ class FieldMockDirective {
   template: `
     <label hLabel #hLabelRef="hLabelRef"></label>
     <label hLabel #hLabelRef="hLabelRef"></label>
-    <label hLabel #hLabelRef="hLabelRef" class="custom" id="custom-id"></label>
-  `,
-  providers: [
-    {
-      provide: FieldDirective,
-      useClass: FieldMockDirective
-    }
-  ]
+    <label hLabel #hLabelRef="hLabelRef"></label>
+  `
 })
 class LabelSpecComponent {}
 
@@ -41,28 +26,14 @@ describe('@headlessng/primitives/label', () => {
       }).createComponent(LabelSpecComponent);
 
       debugs = fixture.debugElement.queryAll(By.directive(LabelDirective));
-      // fixture.autoDetectChanges();
     });
 
     // This test should be run first due to the generation of unique prefixes that are stored in memory.
     it('should generate correct "id" attribute for each element when not passed in input', () => {
-      debugs
-        .map(x => x.nativeElement as HTMLLabelElement)
-        .filter(x => !x.classList.contains('custom'))
-        .forEach((x, i) => {
-          fixture.detectChanges();
-          expect(x.getAttribute('id')).toBe(`h-label-${i}`);
-        });
-    });
-
-    it('should set correct "id" attribute when they are passed to inputs', () => {
-      debugs
-        .map(x => x.nativeElement as HTMLLabelElement)
-        .filter(x => x.classList.contains('custom'))
-        .forEach(x => {
-          fixture.detectChanges();
-          expect(x.getAttribute('id')).toBe('custom-id');
-        });
+      debugs.forEach((x, i) => {
+        fixture.detectChanges();
+        expect(x.nativeElement.getAttribute('id')).toBe(`h-label-${i}`);
+      });
     });
 
     it('should render the correct label elements', () => {
@@ -71,16 +42,6 @@ describe('@headlessng/primitives/label', () => {
 
     it('should forward a references to the directive instance', () => {
       debugs.forEach(x => expect(x.references['hLabelRef']).toBeInstanceOf(LabelDirective));
-    });
-
-    it('should correctly register own reference in the field directive', () => {
-      debugs
-        .map(x => x.references['hLabelRef'] as LabelDirective)
-        .forEach(x => {
-          const registerFn = jest.spyOn(x['_fieldRef'] as FieldDirective, 'registerLabelRef');
-          fixture.detectChanges();
-          expect(registerFn).toHaveBeenCalledWith(x);
-        });
     });
   });
 });